backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-genai==0.3.2
Pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
